Add Render WebAssembly
