--- Medcard.Client/Components/PopupComponent.razor
@@ -1,16 +1,0 @@
-﻿<form method="POST">
-    <div id="popup" class="popup disable">
-        <div class="popup__box">
-            <input placeholder="ФИО" type="text" id="name" name="OwnerName">
-            <input placeholder="Номер телефона" type="text" id="phonenumber" name="PhoneNumber">
-            <input placeholder="Кличка" type="text" id="nick" name="PetName">
-            <input placeholder="Номер чипа" type="text" id="chip" name="ChipNumber">
-            <input placeholder="Возраст" type="text" id="age" name="Age">
-            <input placeholder="Порода" type="text" id="breed" name="Breed">
-
-            <button class="list__more" id="submit">Отправить</button>
-
-        </div>
-        <p></p>
-    </div>
-</form>
