+++ Medcard.Client/Layout/MoreLayout.razor
@@ -1,11 +1,10 @@
-﻿@attribute [StreamRendering(true)]
-@inherits LayoutComponentBase
+﻿
 <head>
-    <link rel="icon" href="/img/ico-logo.png" />
     <link rel="stylesheet" href="css/more.css" />
     <link rel="stylesheet" href="css/normalize.css" />
     <link rel="stylesheet" href="css/site.css" />
+    
 
     <PageTitle>Клиент</PageTitle>
 </head>
-@Body
+
