+++ Medcard.Client/Pages/Clienst.razor
@@ -1,7 +1,8 @@
 ﻿@page "/Client"
-@layout Layout.MoreLayout
 
+@rendermode InteractiveAuto
 
+<MoreLayout></MoreLayout>
 
 <header class="header" id="header">
     <a href="/">
@@ -20,6 +21,7 @@
         <button class="btn-input" id="emailButton" type="button" style="font-size:20px;">Email</button>
     </div>
 </header>
+
 <form method="POST">
     <input type="hidden" name="Id" value="1" />
