+++ Medcard.Client/Program.cs
@@ -1,68 +1,6 @@
-using Medcard.Client.Components;
-using Medcard.DbAccessLayer;
-using Medcard.DbAccessLayer.Dto;
-using Medcard.DbAccessLayer.Entities;
-using Medcard.DbAccessLayer.Interfaces;
-using Medcard.DbAccessLayer.Mapping;
-using Medcard.DbAccessLayer.Repositories;
-using Medcard.Mvc.Abstractions;
-using Medcard.Mvc.Mapping;
-using Medcard.Mvc.Services;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
-var builder = WebApplication.CreateBuilder(args);
+var builder = WebAssemblyHostBuilder.CreateDefault(args);
 
-// Add services to the container.
-builder.Services.AddRazorComponents()
-    .AddInteractiveServerComponents();
 
- 
-builder.Services.AddScoped<IMedcardRepository<OwnerDto>, MedcardRepository>();
-builder.Services.AddScoped<IAuthRepository, AuthRepository>();
-builder.Services.AddScoped<ISearchRepository, SearchRepository>();
-builder.Services.AddSingleton<IHostingServiceMvc, HostingServiceMvc>();
-builder.Services.AddSingleton<IEncrypt, Encrypt>();
-
-// Добавляем распределенный кэш
-builder.Services.AddDistributedMemoryCache();
-
-// Настройки сессий
-builder.Services.AddSession(options =>
-{
-    options.IdleTimeout = TimeSpan.FromMinutes(30);
-    options.Cookie.HttpOnly = true;
-    options.Cookie.IsEssential = true;
-});
-
-builder.Services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationScheme)
-    .AddCookie(options =>
-    {
-        options.LoginPath = "/Authorization/Auth";
-    });
-
-builder.Services.AddHttpContextAccessor();
-builder.Services.AddAutoMapper(typeof(MappingProfile));
-
-// Подключение к базе данных
-builder.Services.AddDbContext<AppDbContext>(options =>
-    options.UseNpgsql(builder.Configuration.GetConnectionString("MedcardConnectionString")));
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (!app.Environment.IsDevelopment())
-{
-    app.UseExceptionHandler("/Error", createScopeForErrors: true);
-    app.UseHsts();
-}
-
-app.UseSession();
-app.UseHttpsRedirection();
-app.UseStaticFiles();
-app.UseAntiforgery();
-
-app.MapRazorComponents<App>()
-    .AddInteractiveServerRenderMode();
-
-app.Run();
+await builder.Build().RunAsync();
