+++ Medcard.Client/_Imports.razor
@@ -1,5 +1,4 @@
-﻿@using System
-@using System.Net.Http
+﻿@using System.Net.Http
 @using System.Net.Http.Json
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
@@ -8,10 +7,6 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 
-@using Medcard.Client.Components.Pages
-@using Medcard.Client.Components.Layout
-@using Medcard.Client.Components
+@using Medcard.Client
+@using Medcard.Client.Layout
 
-@using Medcard.DbAccessLayer
-@using Medcard.DbAccessLayer.Repositories
-@using Medcard.DbAccessLayer.Dto
