+++ Medcard.Server/Components/Layout/MainLayout.razor
@@ -1,0 +1,12 @@
+﻿@inherits LayoutComponentBase
+
+<head>
+    <link rel="stylesheet" href="css/site.css" />
+    <link rel="stylesheet" href="css/normalize.css" />
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css">
+    <PageTitle>Главная</PageTitle>
+</head>
+
+@Body
+
+
