+++ Medcard.Server/Components/Layout/MoreLayout.razor
@@ -1,0 +1,10 @@
+﻿
+<head>
+    <link rel="stylesheet" href="css/more.css" />
+    <link rel="stylesheet" href="css/normalize.css" />
+    <link rel="stylesheet" href="css/site.css" />
+    
+
+    <PageTitle>Клиент</PageTitle>
+</head>
+
