+++ Medcard.Server/Components/_Imports.razor
@@ -1,0 +1,19 @@
+﻿@using System
+@using System.Net.Http
+@using System.Net.Http.Json
+@using Microsoft.AspNetCore.Components.Forms
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
+@using static Microsoft.AspNetCore.Components.Web.RenderMode
+@using Microsoft.AspNetCore.Components.Web.Virtualization
+@using Microsoft.JSInterop
+
+@using Medcard.Server.Components.Pages
+@using Medcard.Server.Components.Layout
+@using Medcard.Server.Components
+
+@using Medcard.DbAccessLayer
+@using Medcard.DbAccessLayer.Repositories
+@using Medcard.DbAccessLayer.Dto
+
+@using Medcard.Client.Pages
