+++ Medcard.Server/Properties/launchSettings.json
