+++ Medcard.Server/appsettings.json
@@ -1,0 +1,10 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning",
+      "Microsoft.EntityFrameworkCore": "Warning" // Добавляем фильтрацию для EF Core
+    }
+  },
+  "AllowedHosts": "*"
+}
